Minor fixes
git-svn-id: svn://bar.ebi.ac.uk/branches/atlas-standalone@16105 2913f559-6b04-0410-9a09-c530ee9f5186

+++ atlas-dao/src/test/java/uk/ac/ebi/gxa/dao/AtlasDAOTestCase.java
@@ -417,7 +417,7 @@ public abstract class AtlasDAOTestCase extends DBTestCase {
         rs.close();
 
         // create an assayid - no oracle id generators here!
-        long assayid = (long) System.currentTimeMillis();
+        long assayid = System.currentTimeMillis();
 
         stmt.executeQuery(
                 "INSERT INTO A2_ASSAY(assayid, accession, experimentid, arraydesignid) " +
@@ -436,7 +436,7 @@ public abstract class AtlasDAOTestCase extends DBTestCase {
         Statement stmt = conn.createStatement();
 
         // create an sampleid - no oracle id generators here!
-        long sampleid = (long) System.currentTimeMillis();
+        long sampleid = System.currentTimeMillis();
 
         stmt.executeQuery(
                 "INSERT INTO A2_SAMPLE(sampleid, accession, channel) " +
